+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Update dictionary data.
 - Update CLI tools deps.
 - Optimized Reflow helper for handling unclosed dialog quote in standalone finalizer.
+- Reflow: Allow commas in title headings when they appear within the first 20 characters.
 
 ---